FEM: import tools, move imported mesh creation def inside new import tools module
