+++ src/Mod/Fem/CMakeLists.txt
@@ -21,6 +21,7 @@ INSTALL(
         importCcxDatResults.py
         importCcxFrdResults.py
         importInpMesh.py
+        importToolsFem.py
         importVTKResults.py
         importZ88Mesh.py
         importZ88O2Results.py
