+++ src/Mod/Fem/importToolsFem.py
@@ -1,0 +1,113 @@
+# ***************************************************************************
+# *                                                                         *
+# *   Copyright (c) 2017 - Bernd Hahnebach <bernd@bimstatik.org>            *
+# *                                                                         *
+# *   This program is free software; you can redistribute it and/or modify  *
+# *   it under the terms of the GNU Lesser General Public License (LGPL)    *
+# *   as published by the Free Software Foundation; either version 2 of     *
+# *   the License, or (at your option) any later version.                   *
+# *   for detail see the LICENCE text file.                                 *
+# *                                                                         *
+# *   This program is distributed in the hope that it will be useful,       *
+# *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+# *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+# *   GNU Library General Public License for more details.                  *
+# *                                                                         *
+# *   You should have received a copy of the GNU Library General Public     *
+# *   License along with this program; if not, write to the Free Software   *
+# *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
+# *   USA                                                                   *
+# *                                                                         *
+# ***************************************************************************
+
+__title__ = "FreeCAD FEM import tools"
+__author__ = "Bernd Hahnebach"
+__url__ = "http://www.freecadweb.org"
+
+## @package importToolsFem
+#  \ingroup FEM
+#  \brief FreeCAD FEM import tools
+
+import FreeCAD
+
+
+def make_femmesh(mesh_data):
+    ''' makes an FreeCAD FEM Mesh object from FEM Mesh data
+    '''
+    import Fem
+    mesh = Fem.FemMesh()
+    m = mesh_data
+    if ('Nodes' in m) and (len(m['Nodes']) > 0):
+        print("Found: nodes")
+        if (('Seg2Elem' in m) or
+           ('Tria3Elem' in m) or
+           ('Tria6Elem' in m) or
+           ('Quad4Elem' in m) or
+           ('Quad8Elem' in m) or
+           ('Tetra4Elem' in m) or
+           ('Tetra10Elem' in m) or
+           ('Penta6Elem' in m) or
+           ('Penta15Elem' in m) or
+           ('Hexa8Elem' in m) or
+           ('Hexa20Elem' in m)):
+
+            nds = m['Nodes']
+            print("Found: elements")
+            for i in nds:
+                n = nds[i]
+                mesh.addNode(n[0], n[1], n[2], i)
+            elms_hexa8 = m['Hexa8Elem']
+            for i in elms_hexa8:
+                e = elms_hexa8[i]
+                mesh.addVolume([e[0], e[1], e[2], e[3], e[4], e[5], e[6], e[7]], i)
+            elms_penta6 = m['Penta6Elem']
+            for i in elms_penta6:
+                e = elms_penta6[i]
+                mesh.addVolume([e[0], e[1], e[2], e[3], e[4], e[5]], i)
+            elms_tetra4 = m['Tetra4Elem']
+            for i in elms_tetra4:
+                e = elms_tetra4[i]
+                mesh.addVolume([e[0], e[1], e[2], e[3]], i)
+            elms_tetra10 = m['Tetra10Elem']
+            for i in elms_tetra10:
+                e = elms_tetra10[i]
+                mesh.addVolume([e[0], e[1], e[2], e[3], e[4], e[5], e[6], e[7], e[8], e[9]], i)
+            elms_penta15 = m['Penta15Elem']
+            for i in elms_penta15:
+                e = elms_penta15[i]
+                mesh.addVolume([e[0], e[1], e[2], e[3], e[4], e[5], e[6], e[7], e[8], e[9],
+                                e[10], e[11], e[12], e[13], e[14]], i)
+            elms_hexa20 = m['Hexa20Elem']
+            for i in elms_hexa20:
+                e = elms_hexa20[i]
+                mesh.addVolume([e[0], e[1], e[2], e[3], e[4], e[5], e[6], e[7], e[8], e[9],
+                                e[10], e[11], e[12], e[13], e[14], e[15], e[16], e[17], e[18], e[19]], i)
+            elms_tria3 = m['Tria3Elem']
+            for i in elms_tria3:
+                e = elms_tria3[i]
+                mesh.addFace([e[0], e[1], e[2]], i)
+            elms_tria6 = m['Tria6Elem']
+            for i in elms_tria6:
+                e = elms_tria6[i]
+                mesh.addFace([e[0], e[1], e[2], e[3], e[4], e[5]], i)
+            elms_quad4 = m['Quad4Elem']
+            for i in elms_quad4:
+                e = elms_quad4[i]
+                mesh.addFace([e[0], e[1], e[2], e[3]], i)
+            elms_quad8 = m['Quad8Elem']
+            for i in elms_quad8:
+                e = elms_quad8[i]
+                mesh.addFace([e[0], e[1], e[2], e[3], e[4], e[5], e[6], e[7]], i)
+            elms_seg2 = m['Seg2Elem']
+            for i in elms_seg2:
+                e = elms_seg2[i]
+                mesh.addEdge(e[0], e[1])
+            print("imported mesh: {} nodes, {} HEXA8, {} PENTA6, {} TETRA4, {} TETRA10, {} PENTA15".format(
+                  len(nds), len(elms_hexa8), len(elms_penta6), len(elms_tetra4), len(elms_tetra10), len(elms_penta15)))
+            print("imported mesh: {} HEXA20, {} TRIA3, {} TRIA6, {} QUAD4, {} QUAD8, {} SEG2".format(
+                  len(elms_hexa20), len(elms_tria3), len(elms_tria6), len(elms_quad4), len(elms_quad8), len(elms_seg2)))
+        else:
+            FreeCAD.Console.PrintError("No Elements found!\n")
+    else:
+        FreeCAD.Console.PrintError("No Nodes found!\n")
+    return mesh
